chore: renamed blocks/settingsBlocks/SettingsBlock.kt to blocks/BaseBlock.kt
